very big commit: initial world gen created, textures correctly applied, dimensions of voxels are customizable

+++ genesim_lab/shaders/chunk.vert
@@ -3,7 +3,7 @@
 #extension GL_ARB_separate_shader_objects : require
 #extension GL_ARB_explicit_uniform_location : require
 
-layout (location = 0) in ivec3 in_position;
+layout (location = 0) in vec3 in_position;
 layout (location = 1) in int voxel_id;
 layout (location = 2) in int face_id;
